refactor(frontend): typo in file name bb-135
