+++ apps/frontend/src/pages/quiz/libs/components/analyzing/analyzing.tsx
