+++ apps/frontend/src/pages/quiz/libs/components/analyzing/styles.module.css
